ivi_tc_defconfig: make most FSes to be modules
Optimize the kernel for Tizen IVI use-case where we use ext4 file-system.
We do not need to compile-in any other file-systems and can keep them to
be kernel modules. This makes the kernel smaller and should presumably
make it boot faster - because the decompression and load time should go
down.

This patch makes vmlinuz 446KiB smaller.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>