packaging: remove junk touch lines
In modern kernel the 'version.h' and 'autoconf.h' live in
'generated/uapi/linux/', not in 'include/linux'. But the silly 'touch' commands
which we had in our spec file _created_ empty 'include/linux/version.h' and
'include/linux/autoconf.h' files, which broke compilation of out-of-tree kernel
modules.

This commit removes the junk 'touch' commands from the spec file.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>