packaging: fix post-uninstall script
When the kernel is upgraded with 'rpm -U', the new kernel is first installed
and creates the /boot/vmlinuz symlink, and the old kernel is then uninstalled
and the ponstun script was incorrectly removing the /boot/vmlinuz symlink,
which made the system unbootable.

The patch fixes the issue, reported in TIVI-609.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>