ACPI / LPSS: add support for Intel BayTrail
Intel BayTrail has almost the same Low Power Subsystem than Lynxpoint with
few differences. Peripherals are clocked with different speeds (typically
lower) and the clock is not always gated. To support this we add
possibility to share a common clock and make clock gating optional.

In addition we rename clk-lpt to clk-lpss to make it clear that it supports
all LPSS peripherals and not tied to Lynxpoint.

(This is based on work by Rafael J. Wysocki <rafael.j.wysocki@intel.com>
and before upstreaming we should get SoB from him).

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>