packaging: remove pre-setup-ivi cruft
Remove old bootloader-related cruft from %post and %postun sections. Now we use
the setup-ivi project which provides nice scripts for adding and removing
kernel entries to the boot-loader configuration files.

At the same time this patch fixes a warning which one could see when building
Tizen IVI image using MIC:

 setup-ivi-bootloader-conf: error: installerfw_restore_env(): can't restore the
 installer framework environment: can't find "/etc/installerfw-environment"

The reason for the warning was that this file is created at the very end, and
it is just not there by the time the kernel package is installed.

It is OK to skip executing "setup-ivi-bootloader-conf" in that case because MIC
will create boot-loaders entries for the kernel later anyway.

Change-Id: Ie60ba5eb558c63da9ce86e08b98ef203a3634ee6
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>