NFS: pn533 : delete timer when polling is complete
When polling is complete, the listen timer is not longer required.
Thus, we delete it. Otherwise, it might be possible that the
expiring timer runs its callback with the mod list already reset.
This would cause a division by zero in pn533_poll_next_mod. A
warning is issued to make sure that we are informed of the condition.

Also, cancel the timer in dep_link_down as it resets the poll mod
count.

Change-Id: Ia3590ec536deeaed3abf36d2c9abf065174d7e33
Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>